[MCH] finalized loading of O2 mapping
The new loadO2Mapping function has been tested with the pre-clustering code and a reference file with input digits, verifying that the pre-clustering produces the same set of pre-clusters when using either readMapping() or loadO2Mapping().
The readMapping() function has been modified to ignore the cathodeID field in the pad unique ID, because the cathodeID is redundant and not available in the O2 mapping. The pre-clustering code has also been modified to equivalently ignore the cathodeID.