[MCH] modified pre-clustering code to make it compatible with O2 mapping
The loadDigits function now ignored the cathodeID field of the digit's uniqueID, because it is not available in the O2 mapping.
The corresponding pad is therefore now searched in both planes of a given detection element.
The code has also been temporarily modified to generate a textual output of the pre-clusters, to allow an easier comparison between the AliROOT and O2 mappings.